+++ src/sns/report/model/reports/params.rs
@@ -1,0 +1,17 @@
+use super::governance::SnsGovernanceParameters;
+use serde::Serialize;
+
+#[derive(Clone, Debug, Eq, PartialEq, Serialize)]
+pub struct SnsParamsReport {
+    pub schema_version: u32,
+    pub network: String,
+    pub sns_wasm_canister_id: String,
+    pub fetched_at: String,
+    pub source_endpoint: String,
+    pub fetched_by: String,
+    pub id: usize,
+    pub name: String,
+    pub root_canister_id: String,
+    pub governance_canister_id: String,
+    pub parameters: SnsGovernanceParameters,
+}